+++ src/Controller/ClientController.php
@@ -7,6 +7,7 @@ use App\Entity\Client;
 use App\Service\UserService;
 use App\Service\ClientService;
 use App\Service\SerializerService;
+use App\Service\SerializerJmsService;
 use Symfony\Contracts\Cache\ItemInterface;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
@@ -22,7 +23,7 @@ class ClientController extends AbstractController
     public function __construct(
         private UserService $userService,
         private ClientService $clientService,
-        private SerializerService $serializerService,
+        private SerializerJmsService $serializerService,
         private ValidatorInterface $validator,
         private TagAwareCacheInterface $cache,
         private TagAwareCacheInterface $cachePool,